+++ src/components/options/center/sections/TabRules/TableRules.vue
@@ -42,7 +42,7 @@
 							<button
 								class="btn btn-xs btn-circle tooltip flex items-center justify-items-center"
 								data-tip="Duplicate"
-								@click="(event) => duplicateRule(event, index)"
+								@click.prevent="(event) => duplicateRule(event, index)"
 							>
 								<DuplicateIcon class="!w-4 !h-4" />
 							</button>
@@ -50,7 +50,7 @@
 							<button
 								class="btn btn-xs btn-circle btn-outline tooltip flex items-center justify-items-center btn-error"
 								data-tip="Delete"
-								@click="(event) => deleteRule(event, index)"
+								@click.prevent="(event) => deleteRule(event, index)"
 							>
 								<DeleteIcon class="!w-4 !h-4" />
 							</button>
@@ -107,8 +107,10 @@ const duplicateRule = async (event: MouseEvent, index: number) => {
 	event.stopPropagation();
 };
 
-const deleteRule = (event: MouseEvent, index: number) => {
-	rulesStore.deleteRule(index);
+const deleteRule = async (event: MouseEvent, index: number) => {
+	if (confirm('Are you sure you want to delete this rule?')) {
+		await rulesStore.deleteRule(index);
+	}
 
 	event.stopPropagation();
 };
